[Authen::Captcha #20432] Broken locking lets images files build up
We recently started using Authen::Captcha (1.23). I noticed that within a few
days of starting to use it, tens of thousands of image files had built up,
which were not being expired, and were nowhere to be found in the database of
codes. This was on busy dual-processor boxes, but I managed to reproduce it on
my own workstation easily:

ab -c3 -n1000 http://some/url/which/generate /captchas

By the time 500 or so requests had passed, I had fairly consistently "lost" a
few images from the database. The debug output showed that older versions of
the data were overwriting more recent versions.

The problem is that one has to obtain a lock, read in the data file, do
everything that you're going to do with it, and then write it back. Instead,
Authen::Captcha was reading in the data file, then closing it and releasing the
lock, then modifying the data (in which time another process could be locking
and reading in the file, obtaining a stale copy of the data), then re-locking
and re-opening it, and so on.

Attached is a solution (tested quite extensively and free of the problem), as
well as the original file, and a diff. You may want to go about the locking
differently - I chose to lock on an external file rather than the database
itself, both being paranoid about the caching issues one can read about in
Locking: The Trouble with fd (man DB_File) (which may not apply here - not
sure), and also believing that it's cleaner/simpler.

Note also that in the attached, Captcha.pm the locks _have_ to be exclusive,
not shared, or the same bug would be reintroduced.

Anyway, I'm hoping that you still maintain this module - I see that there have
been no new releases for a long time.  Please let me know.