Fix #79948: Broken Captcha Image on UI if generated code matches code in db
Reproduced the code from the supplied patch (not directly, patch was for a different Captcha.pm version).
Added tests to certify the bug does not occur.

There is still a problem that cannot be solved with this implementation:
In the very remote case that two different codes generate the same random md5 token, then only the code generated
will give a positive match with check_code().
The older code with the same md5 is overwritten in the database.
To the user it will appear as if she has written the code incorrectly.

This failure is more acceptable than the previous "image not found" one.